requirements.txt
#Django>=5.0,<5.1
gunicorn
python-decouple
psycopg2-binary
dj-database-url
requests
whitenoise
django-allauth[socialaccount]
django-allauth-ui
django-widget-tweaks
slippers
cryptography
stripe